add kotest and start a new test for ConnectedRepository.
move repo tests that need no connection to newreposotorytest
