+++ src/test/kotlin/r2dbcfun/R2dbcTest.kt
@@ -20,7 +20,7 @@ import strikt.assertions.isEqualTo
 /**
  * this is just the r2dbc playground that started this project.
  *
- * @see r2dbcfun.RepositoryTest for api usage.
+ * @see r2dbcfun.RepositoryFunctionalTest for api usage.
  */
 @ExperimentalCoroutinesApi
 class R2dbcTest : JUnit5Minutests {
