+++ src/test/kotlin/r2dbcfun/RepositoryFunctionalTest.kt
@@ -50,7 +50,7 @@ data class User(
 
 @Suppress("SqlResolve")
 @ExperimentalCoroutinesApi
-class RepositoryTest : JUnit5Minutests {
+class RepositoryFunctionalTest : JUnit5Minutests {
     init {
         BlockHound.install()
     }
@@ -320,30 +320,6 @@ class RepositoryTest : JUnit5Minutests {
             }
         }
 
-        context("fail fast error handling") {
-            test("fails fast if PK has more than one field") {
-                data class MismatchPK(override val id: Long, val blah: String) : PK
-                data class Mismatch(val id: MismatchPK)
-                runBlocking {
-                    expectCatching { Repository.create<Mismatch>() }.isFailure()
-                        .isA<RepositoryException>()
-                        .message
-                        .isNotNull()
-                        .contains("PK classes must have a single field of type long")
-                }
-            }
-            test("fails if class contains unsupported fields") {
-                data class Unsupported(val field: String)
-                data class ClassWithUnsupportedType(val id: Long, val unsupported: Unsupported)
-                runBlocking {
-                    expectCatching { Repository.create<ClassWithUnsupportedType>() }.isFailure()
-                        .isA<RepositoryException>()
-                        .message
-                        .isNotNull()
-                        .contains("type Unsupported not supported")
-                }
-            }
-        }
     }
 
     private inline fun <reified T : Any> TestContextBuilder<Connection, Fixture<T>>.makeFixture() {
